Stats restructured and polished
