+++ dashboard/statistics/TroopStatistics/PlotsDeliveredByToopAndDate.md
@@ -6,12 +6,15 @@ layout: home
      imports and prepares any components or utilities needed by the template. -->
 <script setup>
     /*  Import the specific Stats component from its relative path.*/
-    import Stats from '../../../components/statistics/g1/s3.vue';
+    import Stats from '../../../components/statistics/TroopStatistics/PlotsDeliveredByToopAndDate.vue';
 </script>
 
 <!-- Markdown heading that will be rendered as an <h1> element.
      It describes the content: "Cumulative performance per inventory team". -->
-# Abgabe von Ecken durch Trupp nach Datum
+[← Zurück zur Statistikauswahl](/dashboard/statistics/){.vp-button .brand}
+
+
+# Abgabe von Ecken durch Trupp nach Datum  (View)
 
 ::: warning Hinweis
 Diese Statistik(en) sind derzeit ein erster Vorschlag und können noch Fehler enthalten. Weitere Ergänzungen und Korrekturen folgen. Wir bitten um kritische Prüfung und um Vorschläge für weitere Statistiken. Anmerkungen und Vorschläge schicken Sie bitte an bwi-support@thuenen.de.
